backend/requirements.txt
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-cors-headers